modified account

+++ MS/Account-pg/Account/src/main/resources/schema.sql
@@ -2,6 +2,6 @@ create table bank_account
 ( accountid  char(4),
   custid  char(4),
   accounttype  varchar(10),
-  balance long,
+  balance integer,
  currency char(3)
 );
